fix error

+++ CppPrimerPlus/Chapter_6/Exam/Exam_09.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 struct orz{
     string name;
-    int mon;
+    double mon;
 };
 
 int main() {
